+++ src/operators.ts
@@ -910,6 +910,17 @@ export function withRecordType<T extends TypeAssertion>(): (ty: T) => T {
 }
 
 
+export function withMeta<T extends TypeAssertion>(meta: any): (ty: T) => T {
+    return (ty: T) => {
+        const ret: T = ({
+            ...ty,
+            meta,
+        });
+        return ret;
+    };
+}
+
+
 export function withMsg<T extends TypeAssertion>(messages: string | ErrorMessages): (ty: T) => T {
     return (ty: T) => {
         if (ty.kind === 'optional') {